% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_get_f32}
\alias{ggml_get_f32}
\title{Get F32 data}
\usage{
ggml_get_f32(tensor)

ggml_get_f32(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Numeric vector
}
\description{
Get F32 data

Get F32 Data
}
