% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_get_i32}
\alias{ggml_get_i32}
\title{Get I32 Data}
\usage{
ggml_get_i32(tensor)
}
\arguments{
\item{tensor}{Tensor of type GGML_TYPE_I32}
}
\value{
Integer vector
}
\description{
Gets integer data from an I32 tensor (e.g., from ggml_argmax)
}
\examples{
\dontrun{
ctx <- ggml_init(16 * 1024 * 1024)
# Create matrix and find argmax
m <- ggml_new_tensor_2d(ctx, GGML_TYPE_F32, 4, 3)
ggml_set_f32(m, c(1,5,2,3, 4,1,8,2, 2,3,1,9))
idx <- ggml_argmax(ctx, m)
graph <- ggml_build_forward_expand(ctx, idx)
ggml_graph_compute(ctx, graph)
indices <- ggml_get_i32(idx)  # Returns c(1, 2, 3) - indices of max values
ggml_free(ctx)
}
}
