% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_is_contiguous}
\alias{ggml_is_contiguous}
\title{Check if Tensor is Contiguous}
\usage{
ggml_is_contiguous(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Logical
}
\description{
Returns TRUE if tensor data is stored contiguously in memory
}
