% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_is_permuted}
\alias{ggml_is_permuted}
\title{Check if Tensor is Permuted}
\usage{
ggml_is_permuted(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Logical
}
\description{
Returns TRUE if tensor dimensions have been permuted
}
