% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_is_transposed}
\alias{ggml_is_transposed}
\title{Check if Tensor is Transposed}
\usage{
ggml_is_transposed(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Logical
}
\description{
Returns TRUE if tensor has been transposed
}
