% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensors.R
\name{ggml_n_dims}
\alias{ggml_n_dims}
\title{Get Number of Dimensions}
\usage{
ggml_n_dims(tensor)
}
\arguments{
\item{tensor}{Tensor pointer}
}
\value{
Integer number of dimensions (1-4)
}
\description{
Returns the number of dimensions of a tensor
}
