% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_nbytes}
\alias{ggml_nbytes}
\title{Get number of bytes}
\usage{
ggml_nbytes(tensor)

ggml_nbytes(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Number
}
\description{
Get number of bytes

Get Number of Bytes
}
