% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_nelements}
\alias{ggml_nelements}
\title{Get number of elements}
\usage{
ggml_nelements(tensor)

ggml_nelements(tensor)
}
\arguments{
\item{tensor}{Tensor}
}
\value{
Number
}
\description{
Get number of elements

Get Number of Elements
}
