% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{ggml_print_objects}
\alias{ggml_print_objects}
\title{Print Objects in Context}
\usage{
ggml_print_objects(ctx)
}
\arguments{
\item{ctx}{GGML context}
}
\description{
Debug function to print all objects (tensors) in the context
}
