% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R, R/tensors.R
\name{ggml_set_f32}
\alias{ggml_set_f32}
\title{Set F32 data}
\usage{
ggml_set_f32(tensor, data)

ggml_set_f32(tensor, data)
}
\arguments{
\item{tensor}{Tensor}

\item{data}{Numeric vector}
}
\description{
Set F32 data

Set F32 Data
}
