% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggml.R
\docType{package}
\name{ggmlR-package}
\alias{ggmlR}
\alias{ggmlR-package}
\title{ggmlR: GGML Tensor Operations for Machine Learning}
\description{
Provides R bindings to the GGML library for efficient machine learning tensor operations on CPU. Supports basic tensor operations including element-wise addition and multiplication. GGML is a tensor library focused on performance and portability for machine learning.
}
\author{
\strong{Maintainer}: Yuri Baramykov \email{lbsbmsu@mail.ru}

}
\keyword{internal}
